wip reorganzing, remove unneeded validation
